refactored api calls into a weather service
